fix hack of setting UPGRADE to 'true'
Now we check if UPGRADE is defined and only then we attempt to run it.
This avoids having to define it even if there is no need for it and
having to be asked for a password to do nothing.

example: xbps-install -Suy syncs and updates packages, so only one
command is needed and defined in UPDATE.

Before this we set UPGRADE to 'true' in this case so it did nothing. But it
asked for password to do nothing.